+++ src/bson_value.ts
@@ -1,4 +1,5 @@
 import { BSON_MAJOR_VERSION } from './constants';
+import { type InspectFn } from './parser/utils';
 
 /** @public */
 export abstract class BSONValue {
@@ -10,8 +11,20 @@ export abstract class BSONValue {
     return BSON_MAJOR_VERSION;
   }
 
-  /** @public */
-  public abstract inspect(): string;
+  [Symbol.for('nodejs.util.inspect.custom')](
+    depth?: number,
+    options?: unknown,
+    inspect?: InspectFn
+  ): string {
+    return this.inspect(depth, options, inspect);
+  }
+
+  /**
+   * @public
+   * Prints a human-readable string of BSON value information
+   * If invoked manually without node.js.inspect function, this will default to a modified JSON.stringify
+   */
+  public abstract inspect(depth?: number, options?: unknown, inspect?: InspectFn): string;
 
   /** @internal */
   abstract toExtendedJSON(): unknown;